Add page for login failure when authorization is required for reading
When authorization is required to read the wiki, the error message from
a failed GitHub login cannot be displayed.  Instead it will redirect to
GitHub for another OAuth authentication, which can lead to a redirect
loop.  To avoid this, the user is now redirected to a special landing
page to display the login failure.

Note: no changes are made for the case where authorization is not
required to read the wiki.